+++ packages/iov-bns/src/bnsconnector.ts
@@ -1,14 +1,14 @@
 import { ChainConnector } from "@iov/bcp-types";
 
 import { bnsCodec } from "./bnscodec";
-import { Client } from "./client";
+import { BnsConnection } from "./bnsconnection";
 
 /**
  * A helper to connect to a bns-based chain at a given url
  */
 export function bnsConnector(url: string): ChainConnector {
   return {
-    client: () => Client.connect(url),
+    client: () => BnsConnection.establish(url),
     codec: bnsCodec,
   };
 }