+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Breaking changes
 
 * Ed25519KeyringEntry now takes a keypair as an argument in createIdentity()
 * Ed25519SimpleAddressKeyringEntry was removed in favour of Ed25519HdWallet together with HdPaths.simpleAddress
+* Client from @iov/bns was renamed to BnsConnection. The connect() function was renamed to BnsConnection.establish()
 
 ## 0.6.1